+++ src/Array/index.js
@@ -4,11 +4,12 @@
  */
 
 import React from 'react'
-import _ from 'underscore'
 import ArrayContextItem from './ArrayContextItem'
 import {replaceIndexKeys} from '../utility'
 import {propTypes as fieldTypePropTypes} from '../FieldType'
 import generateInputsForKeys from '../utility/generateInputsForKeys'
+import isArray from 'lodash/isArray'
+import without from 'lodash/without'
 
 const propTypes = {
   ...fieldTypePropTypes,
@@ -74,7 +75,7 @@ export default class ArrayComponent extends React.Component {
 
   addItem (itemValue = {}) {
     var newArray = this.props.value
-    if (_.isArray(newArray)) {
+    if (isArray(newArray)) {
       newArray.push(itemValue)
     } else {
       newArray = [itemValue]
@@ -85,7 +86,7 @@ export default class ArrayComponent extends React.Component {
 
   removeItem (index) {
     const value = this.props.value || []
-    var newArray = _.without(value, value[index])
+    var newArray = without(value, value[index])
     this.props.onChange(newArray)
   }