+++ src/utility/clean-fields.js
@@ -1,16 +1,19 @@
-import _ from 'underscore'
+import each from 'lodash/each'
+import last from 'lodash/last'
+import union from 'lodash/union'
+import includes from 'lodash/includes'
 
 export default function (doc, fields) {
   const newDoc = {}
-  _.each(doc, (val, key) => {
+  each(doc, (val, key) => {
     const keys = key.split('.').reduce((memo, value) => {
-      const last = _.last(memo)
-      const pre = last ? last + '.' : ''
-      return _.union(memo, [pre + value])
+      const lastItem = last(memo)
+      const pre = lastItem ? lastItem + '.' : ''
+      return union(memo, [pre + value])
     }, [])
     let contains = false
     keys.forEach(newKey => {
-      if (_.contains(fields, newKey)) {
+      if (includes(fields, newKey)) {
         contains = true
       }
     })