+++ src/utility/generateInputsForKeys.js
@@ -1,14 +1,15 @@
 import Field from '../Field'
-import _ from 'underscore'
 import React from 'react'
+import includes from 'lodash/includes'
+import reject from 'lodash/reject'
 
 export default function (keys, parent = '', schema, omit = []) {
-  const fieldNames = _.reject(keys, (key) => {
+  const fieldNames = reject(keys, (key) => {
     var fullKey = parent ? `${parent}.${key}` : key
     var keySchema = schema.schema(fullKey)
     const options = keySchema.srf || keySchema.mrf
     if (options && options.omit) return true
-    if (_.contains(omit, fullKey)) return true
+    if (includes(omit, fullKey)) return true
   })
   return fieldNames.map((key) => {
     var fullKey = parent ? `${parent}.${key}` : key